# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Social Media Room API - private dual-user digital sanctuary
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a chat tag, with cursor pagination
      parameters:
        - name: tagId
          in: query
          required: true
          schema:
            type: integer
        - name: before
          in: query
          required: false
          schema:
            type: integer
          description: Cursor - get messages before this message ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    patch:
      operationId: updateMessage
      tags: [messages]
      summary: Edit a message's content
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageUpdateInput"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /messages/{messageId}/react:
    post:
      operationId: reactToMessage
      tags: [messages]
      summary: Toggle an emoji reaction on a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated reactions for the message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionToggleResult"
...
components:
  schemas:
...
    ReactionInput:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji
    Message:
      type: object
      properties:
        id:
          type: integer
        tagId:
          type: integer
        senderId:
          type: string
          description: Unique Clerk user ID of the sender
        senderDisplayName:
          type: string
          description: Preferred display name shown in the chat UI for both parties
        senderLoginName:
          type: ["string", "null"]
          description: System login email stored for auditing — never rendered in the UI
        content:
          type: string
        createdAt:
          type: string
          format: date-time
          description: Server-authoritative UTC timestamp
        seenByPartner:
          type: boolean
          description: True once the other user has loaded messages up to and including this one. Once true, the sender can no longer delete it.
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of emoji to list of user IDs who reacted
      required:
        - id
        - tagId
        - senderId
        - senderDisplayName
        - content
        - createdAt
        - seenByPartner
        - reactions
    MessageInput:
      type: object
      properties:
        tagId:
          type: integer
        content:
          type: string
          minLength: 1
      required:
        - tagId
        - content
    MessageUpdateInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
    ReactionToggleResult:
      type: object
      properties:
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - reactions
...