# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Social Media Room API - private dual-user digital sanctuary
 * OpenAPI spec version: 0.1.0
 */
import type { MessageReactions } from './messageReactions';

export interface Message {
  id: number;
  tagId: number;
  /** Unique Clerk user ID of the sender */
  senderId: string;
  /** Preferred display name shown in the chat UI for both parties */
  senderDisplayName: string;
  /**
     * System login email stored for auditing — never rendered in the UI
     * @nullable
     */
  senderLoginName?: string | null;
  content: string;
  /** Server-authoritative UTC timestamp */
  createdAt: Date;
  /** True once the other user has loaded messages up to and including this one. Once true, the sender can no longer delete it. */
  seenByPartner: boolean;
  /** Map of emoji to list of user IDs who reacted */
  reactions: MessageReactions;
}