# lib/api-zod/src/generated/types/reactionToggleResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Social Media Room API - private dual-user digital sanctuary
 * OpenAPI spec version: 0.1.0
 */
import type { ReactionToggleResultReactions } from './reactionToggleResultReactions';

export interface ReactionToggleResult {
  reactions: ReactionToggleResultReactions;
}